requirements.txt
six==1.10.0
Flask==0.12.2
Flask_JsonSchema==0.1.1
flask-cors==3.0.7
bs4==0.0.1
enum34==1.1.6
Flask==0.12.2
Flask_JsonSchema==0.1.1
flask_httpauth==3.2.4
flask-jwt==0.3.2
flask_mongoengine==0.9.3
gunicorn==19.7.1
mock==2.0.0
nltk==3.3
numpy==1.14.0
pandas==0.22.0
passlib==1.7.1
pbr==3.1.1
pep8==1.7.0
piazza-api==0.11.0
progressbar==2.3
pycrypto==2.6.1
pytest==3.3.2
pytest-cov==2.5.1
redis==2.10.6
rq==0.10.0
rq_dashboard==0.3.10
rq_scheduler==0.7.0
scipy==1.0.0
sklearn==0.0
